s3_manager/requirements.txt
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.0.1
